backend/requirements.txt
Flask
gTTS
google-genai
python-dotenv